new preprocessing

+++ L3DAS/preprocessing.py
@@ -94,7 +94,7 @@ if __name__ == '__main__':
                         help='perc split between train and validation sets')
     parser.add_argument('--num_mics', type=int, default=1,
                         help='how many ambisonics mics (1 or 2)')
-    parser.add_argument('--num_data', type=int, default=100,
+    parser.add_argument('--num_data', type=float, default=100,
                         help='how many datapoints per set. 0 means all available data')
     parser.add_argument('--stft_nparseg', type=int, default=256,
                         help='num of stft frames')
